requirements.txt
PyOpenGL==3.1.7
glfw==2.6.2
numpy==2.1.3